[Barry] 修改附魔規則 記錄檔功能

+++ Assets/Scripts/Game/UI/UIPlayerInfo.cs
@@ -87,7 +87,7 @@ namespace Simemes.UI
             UpdateLevel(_currentLevel);
         }
 
-        private void SetExp(int value)
+        private void SetExp(float value)
         {
             _current = value;
             _target = value;
